requirements.txt
matplotlib==3.7.0
mlflow==2.3.0
pandas>=2.0.0
scikit-learn>=1.2.0,<=1.5.2
xgboost>=1.4.2,<3.0.0
psutil>=5.9.0
pyyaml>=6.0
python-dateutil>=2.8.2
pynvml>=11.4.1